net: sched: change block index type from u32 to u64
Following patches in series create special e2e cache block that is used by
cache data path. Lifetime of such block is tied to lifetime of cache object
and not of individual Qdisc. As such, it behaves like shared block (doesn't
store direct pointer to parent Qdisc that can be destroyed before cache
object itself) but needs to not conflict with regular user-created shared
blocks. To support such use-case block index is extended to 64 bits and
special constant TCF_BLOCK_E2E_CACHE if defined outside of 32 bits range.
Blocks with index equal to TCF_BLOCK_E2E_CACHE are not stored in shared
block idr but are treated as shared blocks by all other code.

+++ include/net/sch_generic.h
@@ -425,7 +425,7 @@ struct tcf_block {
 	 */
 	struct mutex lock;
 	struct list_head chain_list;
-	u32 index; /* block index for shared blocks */
+	u64 index; /* block index for shared blocks */
 	u32 classid; /* which class this block belongs to */
 	refcount_t refcnt;
 	struct net *net;
